bcrypt cannot be used in code accessible by server side
