+++ utils/globalFunctions.ts
@@ -2,11 +2,8 @@ import { toast } from '@/components/ui/use-toast';
 import { ToastActionElement } from '@/components/ui/toast';
 import crypto from 'crypto';
 import prisma from '@/prisma';
-import bcrypt from 'bcrypt';
 import { Prisma, GlobalSiteSettings } from '@prisma/client';
 import { defaultSiteSettings } from '@/utils/globalVariables';
-import prismaAdapterHashed from '@/app/api/auth/[...nextauth]/adapters/prismaAdapterHashed';
-import { zCourseRolesType, zSiteRoles } from '@/types/sharedZodTypes';
 
 export function getPublicUrl(): string {
   if (process.env.NEXT_PUBLIC_BASE_URL) {
@@ -138,40 +135,4 @@ export function decrypt(text: string, key?: string) {
   decrypted = Buffer.concat([decrypted, decipher.final()]);
 
   return decrypted.toString();
-}
-
-type CourseMemberInput = {
-  name: string;
-  email: string;
-  role: zCourseRolesType
-  courseId: string;
-  lmsId?: string
-  optionalId?: string
-};
-
-export async function createHashedCourseMember(courseMember: CourseMemberInput) {
-  console.log('Successfully reading site settings from context');
-  const { name, email } = courseMember
-  const hashFunctions = prismaAdapterHashed(prisma);
-  const hashedEmail = await bcrypt.hash(email, 10);
-
-  const existingUser = await hashFunctions.getUserByEmail(email);
-
-  if (!existingUser) {
-    await prisma.user.create({
-      data: {
-        name: name,
-        email: hashedEmail,
-        role: zSiteRoles.enum.user
-      }
-    });
-  }
-
-  const resEnrollment = await prisma.courseMember.create({
-    data: {
-      ...courseMember,
-      email: existingUser?.email ?? hashedEmail
-    }
-  });
-  return resEnrollment;
 }
